fix(savedata): harden backup lifecycle
Make archive creation and record persistence one compensated operation. Handle Windows reparse and case semantics, clean failed restores, and surface alternate restore paths in the UI.

+++ src-tauri/src/database/repository/games_repository.rs
@@ -1063,8 +1063,8 @@ impl GamesRepository {
         db: &DatabaseConnection,
         game_id: i32,
         file_name: &str,
-        backup_time: i32,
-        file_size: i32,
+        backup_time: i64,
+        file_size: i64,
     ) -> Result<i32, DbErr> {
         let savedata_record = savedata::ActiveModel {
             id: NotSet,
